LibGL: Split up GLContext implementation into logical units
This allows for faster rebuilds when the implementation changes.
