+++ Userland/Libraries/LibGL/CMakeLists.txt
@@ -1,8 +1,15 @@
 set(SOURCES
+    ContextParameter.cpp
     GLAPI.cpp
     GLContext.cpp
+    Matrix.cpp
+    Lighting.cpp
+    Lists.cpp
+    Stencil.cpp
     Tex/NameAllocator.cpp
     Tex/Texture2D.cpp
+    Textures.cpp
+    Vertex.cpp
 )
 
 serenity_lib(LibGL gl)
